build: Fix annoying warnings
There are few warnings printed during complation:
- "ignoring #pragma warning [-Wunknown-pragmas]"
It is MSVC-specific pragma setting. Replaced with GCC-specific surrounded
by the compiler-specific macro.
- "comparison between ‘enum _SILABS_PID’ and ‘const CP210X_PARTNUM’ [-Wenum-compare]"
The same warning as one suppressed by the pravious pragma. Just need to add
the "warning push/pop" pragma here.
- "format ‘%lx’ expects argument of type ‘long unsigned int’, but argument 2 has type ‘DWORD'"
Just cast the arguments to unsigned long as the functions name says.
- "variable ‘temp’ set but not used"
Variabe is used only to check whether the passed pointer is valid.
Just used dummy-void-cast operator to silence it up.

+++ smt/src/util.cpp
@@ -27,15 +27,15 @@ void writeStr( std::string s)
 }
 void writeUlong( int val)
 {
-    printf( " %08lx", (DWORD) val);
+    printf( " %08lx", (unsigned long) val);
 }
 void writeUlong( DWORD val)
 {
-    printf( " %08lx", val);
+    printf( " %08lx", (unsigned long) val);
 }
 void writeUlongParm( DWORD val)
 {
-    printf( " { %08lx }", val);
+    printf( " { %08lx }", (unsigned long) val);
 }
 void writeUshort( WORD val)
 {
